Updates for GH comments
