--- client/app/lib/defer.js
@@ -1,8 +1,0 @@
-export default function defer() {
-  const result = {};
-  result.promise = new Promise((resolve, reject) => {
-    result.resolve = resolve;
-    result.reject = reject;
-  });
-  return result;
-}
